Wshorten fixes

+++ include/osmtools/OsmGridRegionTree.h
@@ -202,7 +202,7 @@ public:
 	///this is thread-safe
 	uint32_t push_back(const sserialize::spatial::GeoRegion & p, value_type && value) {
 		std::unique_lock<std::mutex> lck(m_mtx);
-		uint32_t pid = m_regions.size();
+		uint32_t pid = (uint32_t) m_regions.size();
 		this->push_back(&p);
 		m_values.emplace_back(value);
 		return pid;
